refactor: PlayingGameBoard 클래스 추가
게임 진행 중인 기본 게임판과 지뢰찾기 게임판을 관리하고 셀 탐색에 대한 책임을 가진 PlayingGameBoard 클래스 추가
